streamlittest_moonvit/requirements_docker.txt
streamlit
torch
transformers==4.45.0
psycopg2-binary
pgvector
numpy
Pillow
tqdm
